Move the sidebar navigation handlers into useAppNavigation
The entry-click handlers for characters, sets, folders, All Pictures and
the duplicate queue, plus the route pushes that follow them, move out of
App.vue. The direction of travel is what the split is about: reading the
route back into the stores stays with useViewStore and its single route
watcher, while this composable only ever goes the other way - a gesture
updates the selection stores, then pushes the URL that expresses them.

App.vue calls back into the two things it still owns, clearing the search
bar and closing the mobile sidebar, through hooks.

App.vue drops from 2051 to 1735 lines.
